mv locks utils to Balances
