+++ src/contexts/TransferOptions/Utils.ts
@@ -4,36 +4,7 @@
 import type { MaybeAddress } from '@polkadot-cloud/react/types';
 import { unitToPlanck } from '@polkadot-cloud/utils';
 import BigNumber from 'bignumber.js';
-import type { BalanceLock, UnlockChunk } from 'contexts/Balances/types';
 import type { NetworkName } from 'types';
-
-// Gets the total unlocking and unlocked amount.
-export const getUnlocking = (chunks: UnlockChunk[], thisEra: BigNumber) => {
-  let totalUnlocking = new BigNumber(0);
-  let totalUnlocked = new BigNumber(0);
-
-  for (const { value, era } of chunks) {
-    if (thisEra.isGreaterThan(era)) {
-      totalUnlocked = totalUnlocked.plus(value);
-    } else {
-      totalUnlocking = totalUnlocking.plus(value);
-    }
-  }
-  return { totalUnlocking, totalUnlocked };
-};
-
-// Gets the total locked amount from an account's locks.
-export const getLocked = (locks: BalanceLock[]) =>
-  locks?.reduce((prev, { amount }) => prev.plus(amount), new BigNumber(0)) ||
-  new BigNumber(0);
-
-// Gets the largest lock balance, dictating the total amount of unavailable funds from locks.
-export const getMaxLock = (locks: BalanceLock[]) =>
-  locks.reduce(
-    (prev, current) =>
-      prev.amount.isGreaterThan(current.amount) ? prev : current,
-    { amount: new BigNumber(0) }
-  )?.amount || new BigNumber(0);
 
 // Get the local storage record for an account reserve balance.
 export const getLocalFeeReserve = (
